chore(masthead): introduce composite (#3582)
This change extracts out the rendering portions of
`<dds-masthead-container>` and `<dds-dotcom-shell-container>` to
`<dds-masthead-composite>` and `<dds-dotcom-shell-composite>`,
respectively.

`<dds-masthead-composite>` and `<dds-dotcom-shell-composite>` provides
our users with the choice of Redux store and its reducer.
