+++ packages/web-components/src/components/dotcom-shell/dotcom-shell-container.scss
@@ -7,6 +7,7 @@
 
 @import '../../globals/scss/vars';
 
-:host(#{$dds-prefix}-dotcom-shell-container) {
+:host(#{$dds-prefix}-dotcom-shell-container),
+:host(#{$dds-prefix}-dotcom-shell-composite) {
   display: contents;
 }
