--- packages/web-components/src/components/dotcom-shell/index.ts
@@ -1,14 +1,0 @@
-/**
- * @license
- *
- * Copyright IBM Corp. 2020
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-import { Action, Reducer } from 'redux';
-import './dotcom-shell-container';
-import { reducers, store } from '../masthead/masthead-container';
-
-store.replaceReducer(reducers as Reducer<unknown, Action<any>>);
