+++ packages/web-components/src/globals/services-store/store.ts
@@ -10,6 +10,7 @@
 import { createStore, applyMiddleware } from 'redux';
 import thunkMiddleware from 'redux-thunk';
 import { createLogger } from 'redux-logger';
+import reducers from './reducers';
 
 const middlewares = [thunkMiddleware];
 
@@ -17,5 +18,5 @@ if (process.env.NODE_ENV === 'development') {
   middlewares.push(createLogger());
 }
 
-const store = createStore(state => state, {}, applyMiddleware(...middlewares));
+const store = createStore(reducers, {}, applyMiddleware(...middlewares));
 export default store;
